requirements.txt
django
python-dotenv
django-phonenumber-field[phonenumbers]
psycopg2
gunicorn
Pillow
djangorestframework